Add an alternate formatting (sandwich) mode
This adds a new sandwich mode and a new command /rasw.
If the user prefers to have the reply sandwiched between usernames like
'@screen_name witty_reply @mentioned' instead of the default
'@screen_name @mentioned witty_reply', he can use the new /rasw command.
This method of replying-to-all is quite common in the twittersphere.

Signed-off-by: Thameera Senanayaka <thameera123@gmail.com>